Parse the number of samples as a double.
The number of samples can grow very large depending on the hardware so a
double should be used to store them. Samples larger than
Integer.MAX_VALUE should be represented in scientific notation.

Reviewed-on: https://git.eclipse.org/r/11094
Tested-by: Hudson CI
IP-Clean: Camilo Bernal <cabernal@redhat.com>
Tested-by: Camilo Bernal <cabernal@redhat.com>

+++ perf/org.eclipse.linuxtools.perf/src/org/eclipse/linuxtools/internal/perf/PerfCore.java
@@ -439,7 +439,7 @@ public class PerfCore {
 		float percent;
 
 		Process p = null;
-		int samples;
+		double samples;
 		String comm,dso,symbol;
 		boolean kernelFlag;
 		PMEvent currentEvent = null;
@@ -476,7 +476,7 @@ public class PerfCore {
 						continue;
 					}
 					percent = Float.parseFloat(items[0]); //percent column
-					samples = Integer.parseInt(items[1].trim()); //samples column
+					samples = Double.parseDouble(items[1].trim()); //samples column
 					comm = items[2].trim(); //command column
 					dso = items[3].trim(); //dso column
 					symbol = items[4].trim(); //symbol column
